Add test case
Added a test case to test the `run` function when all action inputs are present. This test also includes createCommentInput that represents the issue comment message when these inputs are present the the message input variable is formatted into the issue comment body. With this we test 100% of the `run` function.